requirements.txt
# FastAPI and Server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
starlette>=0.27.0
itsdangerous>=2.1.0

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# OpenAI (LLM and Embeddings)
openai>=1.0.0

# Vector Database
pinecone>=5.0.0

# Environment and Configuration
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic[email]>=2.0.0
email-validator>=2.0.0

# Document Processing
pypdf2>=3.0.0
pdfplumber>=0.10.0
langchain>=0.1.0
langchain-text-splitters>=0.0.1

# Web Scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
requests>=2.31.0

# Templates and Static Files
jinja2>=3.1.0
aiofiles>=23.2.0

# File Type Detection (for input validation)
python-magic>=0.4.27

# Authentication
PyJWT>=2.8.0
bcrypt>=4.1.0

